feat(app version widget): factors it out into its own source file PE-4533
