+++ lib/components/app_version_widget.dart
@@ -1,0 +1,37 @@
+import 'package:ardrive/utils/app_localizations_wrapper.dart';
+import 'package:ardrive/utils/logger/logger.dart';
+import 'package:ardrive_ui/ardrive_ui.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+
+class AppVersionWidget extends StatelessWidget {
+  const AppVersionWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder(
+      future: PackageInfo.fromPlatform(),
+      builder: (BuildContext context, AsyncSnapshot<PackageInfo> snapshot) {
+        final info = snapshot.data;
+        if (info == null) {
+          logger.d('PackageInfo is null');
+          return const SizedBox(
+            height: 32,
+            width: 32,
+          );
+        }
+        final literalVersion =
+            kIsWeb ? info.version : '${info.version}+${info.buildNumber}';
+        logger.d('Version: $literalVersion');
+        return Text(
+          appLocalizationsOf(context).appVersion(literalVersion),
+          style: ArDriveTypography.body.buttonNormalRegular(
+            color: Colors.grey,
+          ),
+          textAlign: TextAlign.center,
+        );
+      },
+    );
+  }
+}
