+++ lib/components/side_bar.dart
@@ -1,6 +1,7 @@
 import 'package:ardrive/blocs/drive_detail/drive_detail_cubit.dart';
 import 'package:ardrive/blocs/drives/drives_cubit.dart';
 import 'package:ardrive/blocs/profile/profile_cubit.dart';
+import 'package:ardrive/components/app_version_widget.dart';
 import 'package:ardrive/components/new_button/new_button.dart';
 import 'package:ardrive/components/theme_switcher.dart';
 import 'package:ardrive/misc/resources.dart';
@@ -15,7 +16,6 @@ import 'package:ardrive_ui/ardrive_ui.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:package_info_plus/package_info_plus.dart';
 import 'package:responsive_builder/responsive_builder.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -714,35 +714,6 @@ class HelpButton extends StatelessWidget {
       icon: ArDriveIcons.question(),
       onPressed: () {
         openUrl(url: Resources.helpLink);
-      },
-    );
-  }
-}
-
-class AppVersionWidget extends StatelessWidget {
-  const AppVersionWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder(
-      future: PackageInfo.fromPlatform(),
-      builder: (BuildContext context, AsyncSnapshot<PackageInfo> snapshot) {
-        final info = snapshot.data;
-        if (info == null) {
-          return const SizedBox(
-            height: 32,
-            width: 32,
-          );
-        }
-        final literalVersion =
-            kIsWeb ? info.version : '${info.version}+${info.buildNumber}';
-        return Text(
-          appLocalizationsOf(context).appVersion(literalVersion),
-          style: ArDriveTypography.body.buttonNormalRegular(
-            color: Colors.grey,
-          ),
-          textAlign: TextAlign.center,
-        );
       },
     );
   }
